Tiki Trap Changes
Restructured the tiki trap such that it fires projectiles instead of particle effects. Designers can now add a variable amount of weapons to the actual tiki totem and specify when to start shooting a projectile and the cool-down or length of time between each shot. Added a tiki trap to Level 1 as a demonstration.