Super Shiny Update of Awesomeness
This commit contains a shitload of new stuff.  First, 90% of all sound effects for level 1 are added.  The exception is that of the hulk and the kushling, whose sound effects don't work... yet.  Secondly, the moving platform is fixed ala wizardy of just setting it's ResolveCollision to none (stupid built-in collision resolutions...).  The hulk's ranged attack was also changed to allow tracking of the player, instead of some arbitrary length.  However, this isn't 100% operational as a small bug that doubles the velocity of each subsequent projectile is hiding in there somewhere.

Also, the TorqueConsole has been integrated into the latest code.  This exposes debug statements that should be useful for Matt and Andy while building/testing levels.  It also has the feature of running custom routines to affect the game at runtime.  Currently, the only routines exposed to this are HealPlayer(), WarpToCheckpoint(), and LoadLevel().  LoadLevel() is the only one that is still in an unstable state, but that's because of the underlying level loading/unloading framework of TorqueX which I should have figured out by next commit.