Loading Screen Commit
Added loading screen after player initially starts the first level.  This lets
the player know that the game is in fact loading and not frozen.  Also added
golden crystal pick-up component that is added to Amanda's inventory when
collided with (picked-up).  Mobs can also now drop a predetermined amount of
crystals, whether they be health or generic crystals.