Oooh!  Shiny things!!
Added crystal art assets, sound effects for the crystal pickups, and modified
the collectible components to play a sfx and particle when the collectible is
picked up.  As a note, all child classes of CollectibleComponent now have a
SpawnedParticle that contains a particle to create and play when the collectible
is picked up.